+++ tests/player_death_model.h
@@ -13,8 +13,8 @@ struct test_player_death_state {
 	const uint8_t *current_name;
 	size_t current_name_length;
 	struct yt_player victim;
-	int old_ports_owned;
-	int matched_ports;
+	uint16_t old_ports_owned;
+	uint16_t matched_ports;
 	uint8_t victim_name[YT_TEXT_FIELD_SIZE];
 	size_t victim_name_length;
 	bool complete;
